:art: reformat routes
