+++ api/__init__.py
@@ -1,3 +1,3 @@
 from api.assessments.routes import AssessmentsView  # noqa
-from api.assessments.routes import ScoresJustificationsView  # noqa
-from api.assessments.routes import SubCriteriaView  # noqa
+from api.scores_justifications.routes import ScoresJustificationsView  # noqa
+from api.subcriterias.routes import SubCriteriaView  # noqa
